deeptechfinder/requirements.txt
# DeepTechFinder Patent Analytics Requirements

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0
urllib3>=1.26.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# PDF generation
reportlab>=4.0.0

# Progress bars and CLI
tqdm>=4.65.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Documentation
markdown>=3.4.0